backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
sqlalchemy>=2.0.22
psycopg2-binary>=2.9.9
pgvector>=0.2.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
